Creates a formatted DT::datatable

Creates a formatted DT::datatable with many adjustable formatting options. Designed to be called within v() or v_shiny_internal for easier formatting adjustments and server-side processing.

create_v_datatable(
  .df,
  .subject_col = NULL,
  .freeze_cols = NULL,
  .digits = 3,
  .show_filters = FALSE,
  .show_labels = TRUE,
  .wrap_labels = FALSE,
  .trunc_labels = FALSE,
  .trunc_length = 20,
  .ft_size = 9,
  .scroll_y = 400,
  .subj_contrast = FALSE
)

Arguments

.df

A dataframe that you want to process and view.

.subject_col

A character string specifying the subject column. If this column is present in the dataframe, it will be used to group rows in the datatable.

.freeze_cols

A character vector specifying columns to freeze when scrolling horizontally. .subject_col will automatically be appended to this list (i.e. .subject_col is always frozen).

.digits

number of digits to round numeric columns to. Set to NULL to prevent rounding.

.show_filters

Logical (TRUE/FALSE). If TRUE, show column filters.

.show_labels

Logical (TRUE/FALSE). If TRUE, show label attributes associated with the columns.

.wrap_labels

Logical (TRUE/FALSE). If TRUE, wrap column labels to have multuple lines.

.trunc_labels

Logical (TRUE/FALSE). If TRUE, truncate the column label attributes.

.trunc_length

Numeric. Maximum width of column labels. Only relevant if .trunc_labels = TRUE.

.ft_size

Numeric. Base font size of the table. Other labels scale off this value.

.scroll_y

Numeric. Height of the table in pixels. Passed to scrollY option of DT::datatable.

.subj_contrast

Logical (TRUE/FALSE). If TRUE, increase the color contrast between unique .subject_col groupings.

Value

An instance of DT::datatable class with the processed dataframe.

Details

If .subject_col is NULL, the column names "USUBJID" and "ID" will be searched and set if present. If .subject_col is present in the dataframe, it will be frozen and alternate in color.

Columns with less than 20 unique values are converted to factors. This can be helpful for filtering, which will show up as the shiny equivalent of selectInput, as opposed to the default textInput for character columns.

Columns with label attributes will be appended to column headers as a new line.

You can drag and drop columns to move them around in the table. If a column is dragged to a frozen column's location, the new column will be frozen instead.