Write an NMTRAN data set and corresponding meta data

nm_write is deprecated as of mrgda version 0.7.0. Please use write_derived instead.

This function will take a nonmem ready data frame in R and write it out to a nonmem ready csv. It also creates a meta data folder, storing the xpt file along with other useful information.

nm_write(
  .data,
  .spec,
  .file,
  .prev_file = NULL,
  .compare_from_svn = TRUE,
  .return_base_compare = FALSE
)

Arguments

.data

a data frame

.spec

a yspec object

.file

csv file name to write out to (including path)

.prev_file

csv file name of previous version (defaults to .file)

.compare_from_svn

logical. Should the data comparison be done on the latest svn version? (If not, local version is used)

.return_base_compare

logical. Should the two current and previous versions of the datasets be returned?

Examples

if (FALSE) {
nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))
nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")
nm_write(.data = nm, .spec = nm_spec, .file = "data/derived/pk.csv")
}