Source data query

Use this to search for a character string across every source data domain including the name, label and contents of each domain.

query_src_dir(.src_directory, .string, .file_types = "detect")

Arguments

.src_directory

file path to the source data directory

.string

string to search for

.file_types

Type of files being read in (e.g. 'sas7bat'). The default ('detect') will determine file type based on the most occurring file type in .path.

Examples

path <- system.file("example-sdtm", package = "mrgda")

# Summarize source directory
query_src_dir(.src_directory = path, .string = "RACE")
#> ℹ Detected file type = 'xpt'
#> Reading in ae.xpt (6.88 KB)
#> ✔ /data/mrgda/inst/example-sdtm/ae.xpt
#> Reading in ae.xpt (6.88 KB)
#> Reading in dm.xpt (3.76 KB)
#> ✔ /data/mrgda/inst/example-sdtm/dm.xpt
#> Reading in dm.xpt (3.76 KB)
#> Reading in eg.xpt (14 KB)
#> ✔ /data/mrgda/inst/example-sdtm/eg.xpt
#> Reading in eg.xpt (14 KB)
#> Reading in lb.xpt (24.24 KB)
#> ✔ /data/mrgda/inst/example-sdtm/lb.xpt
#> Reading in lb.xpt (24.24 KB)
#> Reading in mh.xpt (5.12 KB)
#> ✔ /data/mrgda/inst/example-sdtm/mh.xpt
#> Reading in mh.xpt (5.12 KB)
#> Reading in pe.xpt (12.8 KB)
#> ✔ /data/mrgda/inst/example-sdtm/pe.xpt
#> Reading in pe.xpt (12.8 KB)
#> Reading in vs.xpt (14.08 KB)
#> ✔ /data/mrgda/inst/example-sdtm/vs.xpt
#> Reading in vs.xpt (14.08 KB)
#> ┌ read_src_dir Summary ────────────────────────┐
#> │                                              │
#> │   Number of domains successfully loaded: 7   │
#> │   Number of domains that failed to load: 0   │
#> │                                              │
#> └──────────────────────────────────────────────┘
#> # A tibble: 5 × 4
#>   DOMAIN            COLUMN       MATCH             MATCHING_ROWS             
#>   <chr>             <chr>        <chr>             <chr>                     
#> 1 lb                LBORRES      TRACE             20, 80                    
#> 2 lb                LBSTNRC      NEGATIVE TO TRACE 19, 20, 38, 58, 59, 80, 81
#> 3 lb                LBSTRESC     TRACE             20, 80                    
#> 4 mrgda_labels (dm) COLUMN_LABEL RACE_LABEL        30                        
#> 5 mrgda_labels (dm) COLUMN_NAME  RACE              30