Query source data domains for a topic

Search a source list for a string across column names, column labels, column values so you can quickly locate data of interest.

query_src_list(.src_list, .string, .ignore_case = TRUE)

Arguments

.src_list

A source list created using read_src_dir.

.string

string to search for (not case sensitive by default)

.ignore_case

logical. Should the search not be case sensitive? Defaults to TRUE

Examples

path <- system.file("example-sdtm", package = "mrgda")
src_list <- read_src_dir(path, .file_types = "xpt")
#> ℹ User specified file type = 'xpt'
#> Reading in ae.xpt (6.88 KB)
#> ✔ /Users/andersone/Library/R/arm64/4.5/library/mrgda/example-sdtm/ae.xpt
#> Reading in ae.xpt (6.88 KB)
#> Reading in dm.xpt (3.76 KB)
#> ✔ /Users/andersone/Library/R/arm64/4.5/library/mrgda/example-sdtm/dm.xpt
#> Reading in dm.xpt (3.76 KB)
#> Reading in eg.xpt (14 KB)
#> ✔ /Users/andersone/Library/R/arm64/4.5/library/mrgda/example-sdtm/eg.xpt
#> Reading in eg.xpt (14 KB)
#> Reading in lb.xpt (24.24 KB)
#> ✔ /Users/andersone/Library/R/arm64/4.5/library/mrgda/example-sdtm/lb.xpt
#> Reading in lb.xpt (24.24 KB)
#> Reading in mh.xpt (5.12 KB)
#> ✔ /Users/andersone/Library/R/arm64/4.5/library/mrgda/example-sdtm/mh.xpt
#> Reading in mh.xpt (5.12 KB)
#> Reading in pe.xpt (12.8 KB)
#> ✔ /Users/andersone/Library/R/arm64/4.5/library/mrgda/example-sdtm/pe.xpt
#> Reading in pe.xpt (12.8 KB)
#> Reading in vs.xpt (14.08 KB)
#> ✔ /Users/andersone/Library/R/arm64/4.5/library/mrgda/example-sdtm/vs.xpt
#> Reading in vs.xpt (14.08 KB)
#> ┌ read_src_dir Summary ────────────────────────┐
#> │                                              │
#> │   Number of domains successfully loaded: 7   │
#> │   Number of domains that failed to load: 0   │
#> │                                              │
#> └──────────────────────────────────────────────┘

query_src_list(src_list, .string = "RACE")
#> # A tibble: 5 × 4
#>   DOMAIN COLUMN   MATCH_TYPE VALUE            
#>   <chr>  <chr>    <chr>      <chr>            
#> 1 dm     RACE     column     NA               
#> 2 dm     RACE     label      RACE_LABEL       
#> 3 lb     LBORRES  value      TRACE            
#> 4 lb     LBSTNRC  value      NEGATIVE TO TRACE
#> 5 lb     LBSTRESC value      TRACE            
query_src_list(src_list, .string = "F")
#> # A tibble: 41 × 4
#>    DOMAIN COLUMN   MATCH_TYPE VALUE                   
#>    <chr>  <chr>    <chr>      <chr>                   
#>  1 ae     AEENRF   column     NA                      
#>  2 ae     AEMODIFY column     NA                      
#>  3 ae     AEENRF   label      AEENRF_LABEL            
#>  4 ae     AEMODIFY label      AEMODIFY_LABEL          
#>  5 ae     AEDECOD  value      Fatigue                 
#>  6 ae     AEENRF   value      AFTER                   
#>  7 ae     AETERM   value      LEFT KNEE PAIN WORSENING
#>  8 dm     RFENDTC  column     NA                      
#>  9 dm     RFSTDTC  column     NA                      
#> 10 dm     RFENDTC  label      RFENDTC_LABEL           
#> # ℹ 31 more rows