Run application in background for single user

Use callr::r_bg() to launch the app in a separate process, returning once port on host is active.

run_app_bg(
  func,
  args,
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port")
)

Arguments

func

Function passed as the func argument to callr::r_bg(). This must meet the following requirements:

launch a Shiny app

accept host and port arguments and instruct Shiny to use those

call devtools::load_all() with the value of the MRGDA_SHINY_DEV_LOAD_PATH environment variable if it is a non-empty string.

args

Argument passed as args argument to callr::r_bg(). This should not include "host" or "port"; those will be added by this function. Must be a named list.

host, port

Host and port to serve the Shiny app on. If port isn't specified, it is randomly selected from the valid range of dynamic ports.

Value

The callr process (invisibly).

Details

Make sure to set Sys.setenv('MRGDA_SHINY_DEV_LOAD_PATH' = rprojroot::find_rstudio_root_file()) if you are in a development setting. This will allow package functions to be accessible in the background process, which is not an issue when the package is installed.

Examples

if (FALSE) {

shiny_func <- function(
   df,
   filter = TRUE,
   host = NULL,
   port = NULL
){

 # For Development Environment, must load the package
 load_path <- Sys.getenv("MRGDA_SHINY_DEV_LOAD_PATH")
 if (nzchar(load_path)) {
   message("Loading ", load_path)
   devtools::load_all(load_path)
 }

 # other setup...

 ui <- fluidPage
 server <- function(input, output, session) {}

 # Passes the host and port to the shinyApp object
 app <- shiny::shinyApp(ui = ui, server = server,
                 options = list(host = host, port = port))

 # Function launches the app
 shiny::runApp(app)
}

# Launching app in background
args <- list(df = mtcars, filter = TRUE)
run_app_bg(shiny_func, args = args)

}